requirements.txt
python-docx>=1.2.0
lxml>=3.1.0
typing-extensions>=4.9.0
